requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-chess==1.999
pydantic==2.9.0
